Load configuration and service data from system level directories.
We check three locations for data, in order from least-specific to
most-specific: a system directory, the user's home directory, and the
current directory.

First, we'll load configuration data and overwrite data from more
generic locations with the more specific locations.  Next, we'll load
service data from the most specific location to the most generic
location, stopping as soon as we find service data.