Prefer non-expired keys when picking keys automagically.
Previously, if the user didn't specify a key id, we picked the first
secret key we found, regardless of whether that key was expired or
not.  That behavior was inconvenient, because it could select
"invalid" (for some definitions of invalid) keys.

This behavior has been fixed by picking the first non-expired key we
find instead.  If we do not find a non-expired key, we select the last
key we checked instead (usually the one most recently added to the
keyring).